fix: different tables scroll respectively
